+++ calc-mbr.h
@@ -7,14 +7,14 @@
 struct MbrProp
 {
     int size_x, size_y, iters;
-    double x0, y0, scale;
+    float x0, y0, scale;
 };
 
 clock_t calc_set_unopt(MbrProp* p, std::vector<std::vector<int>>* set);
 clock_t calc_set_unr(MbrProp* p, std::vector<std::vector<int>>* set);
 extern clock_t (*calc_set)(MbrProp* p, std::vector<std::vector<int>>* set);
 
-double* time_arr(MbrProp* p, int n);
+float* time_arr(MbrProp* p, int n);
 int measure_mbr_time(MbrProp* p, int n, const char* file_path);
 
 #endif // CALC_MBR_H
